update 'config' dir to match name for ~/.config
The 'src/config_files/configs' directory had that trailing 's' at the
end of it, which was creating the symlink as '~/.configs' instead of
creating the symlink as '~/.config'.

The install script can now just loop through the files in 'config_files'
and just prepend a '.' to the symlinked file name.